Add util.ppy and moved function
