+++ src/management_server/node_allocator/terraform/private_cloud/log.log
@@ -1,0 +1,22 @@
+2023-04-28 15:13:33,676 - utils - ERROR - ╷
+│ Error: Backend initialization required: please run "terraform init"
+│ 
+│ Reason: Backend configuration block has changed
+│ 
+│ The "backend" is the interface that Terraform uses to store state,
+│ perform operations, etc. If this message is showing up, it means that the
+│ Terraform configuration you're using is using a custom configuration for
+│ the Terraform backend.
+│ 
+│ Changes to backend configurations require reinitialization. This allows
+│ Terraform to set up the new configuration, copy existing state, etc. Please
+│ run
+│ "terraform init" with either the "-reconfigure" or "-migrate-state" flags
+│ to
+│ use the current configuration.
+│ 
+│ If the change reason above is incorrect, please verify your configuration
+│ hasn't changed and try again. At this point, no changes to your existing
+│ configuration or state have been made.
+╵
+
